--- a/.eslintrc
+++ b/.eslintrc
@@ -16,7 +16,6 @@
   },
   "rules": {
     "curly": "error",
-    "@typescript-eslint/indent": "off",
     "@typescript-eslint/no-non-null-assertion": "off",
     "@typescript-eslint/no-empty-function": "off",
     "@typescript-eslint/ban-ts-comment": "warn",
@@ -28,7 +27,7 @@
     "react-hooks/exhaustive-deps": "warn",
     "react/display-name": "warn",
     "react/prop-types": "off",
-    "no-console": ["error", { "allow": ["warn", "error"] }],
+    "no-console": ["error"],
     "simple-import-sort/imports": [
       "error",
       {