--- a/app/src/test.tsx
+++ b/app/src/test.tsx
@@ -1,81 +1,52 @@
+// @ts-nocheck
+
 import * as React from 'react';
 import { useFieldArray, useForm } from 'react-hook-form';
 import { Control } from '../../src/types';
+import { Controller } from '../../src';
 
 type FormValues = {
-  nest: {
-    test: {
-      value: string;
-      nestedArray: {
-        value: string;
-      }[];
-    }[];
-  };
+  names: {
+    name: string;
+  }[];
 };
-const ChildComponent = ({
-  index,
-  control,
-}: {
-  control: Control<FormValues>;
-  index: number;
-}) => {
-  const { fields } = useFieldArray<FormValues>({
-    name: `nest.test.${index}.nestedArray` as const,
-    control,
-  });
 
-  return (
-    <div>
-      {fields.map((item, i) => (
-        <input
-          key={item.id}
-          {...control.register(
-            `nest.test.${index}.nestedArray.${i}.value` as const,
-          )}
-        />
-      ))}
-    </div>
-  );
-};
+const output: object[] = [];
 
 const Component = () => {
-  const { register, control } = useForm({
+  const { control, handleSubmit, watch } = useForm<FormValues>({
     defaultValues: {
-      nest: {
-        test: [
-          { value: '1', nestedArray: [{ value: '2' }] },
-          { value: '3', nestedArray: [{ value: '4' }] },
-        ],
-      },
+      names: [],
     },
   });
-  const { fields, remove, append } = useFieldArray({
-    name: 'nest.test',
+  const { fields, append } = useFieldArray({
     control,
+    name: 'names',
   });
 
+  const handleAddElement = () => {
+    append({ name: 'test' });
+  };
+
+  console.log(watch());
+
   return (
-    <div>
-      {fields.map((item, i) => (
-        <div key={item.id}>
-          <input {...register(`nest.test.${i}.value` as const)} />
-
-          <ChildComponent control={control} index={i} />
-
-          <button
-            type={'button'}
-            onClick={() => remove(i)}
-            data-testid={item.value}
-          >
-            remove
-          </button>
-        </div>
-      ))}
-
-      <button type={'button'} onClick={() => append({ value: 'test' })}>
-        append
+    <form onSubmit={handleSubmit(() => {})}>
+      {fields.map((item, index) => {
+        return (
+          <div key={item.id}>
+            <Controller
+              control={control}
+              name={`names.${index}.name` as const}
+              render={({ field }) => <input {...field} />}
+            />
+          </div>
+        );
+      })}
+      <button type="button" onClick={handleAddElement}>
+        Append
       </button>
-    </div>
+    </form>
   );
 };