--- a/app/src/test.tsx
+++ b/app/src/test.tsx
@@ -1,82 +1,52 @@
+// @ts-nocheck
 import * as React from 'react';
 import { useFieldArray, useForm } from 'react-hook-form';
 import { Control } from '../../src/types';
+import { Controller } from '../../src';
 
-type FormValues = {
-  nest: {
-    test: {
-      value: string;
-      nestedArray: {
-        value: string;
-      }[];
-    }[];
-  };
-};
-const ChildComponent = ({
-  index,
-  control,
-}: {
-  control: Control<FormValues>;
-  index: number;
-}) => {
-  const { fields } = useFieldArray<FormValues>({
-    name: `nest.test.${index}.nestedArray` as const,
+const App = () => {
+  const {
     control,
-  });
-
-  return (
-    <div>
-      {fields.map((item, i) => (
-        <input
-          key={item.id}
-          {...control.register(
-            `nest.test.${index}.nestedArray.${i}.value` as const,
-          )}
-        />
-      ))}
-    </div>
-  );
-};
-
-const Component = () => {
-  const { register, control } = useForm({
+    handleSubmit,
+    formState: { errors },
+  } = useForm({
     defaultValues: {
-      nest: {
-        test: [
-          { value: '1', nestedArray: [{ value: '2' }] },
-          { value: '3', nestedArray: [{ value: '4' }] },
-        ],
-      },
+      test: [{ firstName: 'test' }],
     },
   });
-  const { fields, remove, append } = useFieldArray({
-    name: 'nest.test',
+  const { fields, prepend } = useFieldArray({
     control,
+    name: 'test',
   });
+  console.log('errors', errors);
 
   return (
-    <div>
-      {fields.map((item, i) => (
-        <div key={item.id}>
-          <input {...register(`nest.test.${i}.value` as const)} />
-
-          <ChildComponent control={control} index={i} />
-
-          <button
-            type={'button'}
-            onClick={() => remove(i)}
-            data-testid={item.value}
-          >
-            remove
-          </button>
-        </div>
-      ))}
-
-      <button type={'button'} onClick={() => append({ value: 'test' })}>
-        append
+    <form onSubmit={handleSubmit(() => {})}>
+      {fields.map((field, index) => {
+        return (
+          <div key={field.id}>
+            <Controller
+              control={control}
+              render={({ field }) => <input {...field} />}
+              name={`test.${index}.firstName`}
+              rules={{ required: true }}
+            />
+          </div>
+        );
+      })}
+      <button
+        type="button"
+        onClick={() =>
+          prepend({
+            firstName: '',
+          })
+        }
+      >
+        prepend
       </button>
-    </div>
+      <button>submit</button>
+    </form>
   );
 };
 
-export default Component;
+export default App;