--- a/app/src/test.tsx
+++ b/app/src/test.tsx
@@ -1,52 +1,82 @@
-// @ts-nocheck
 import * as React from 'react';
 import { useFieldArray, useForm } from 'react-hook-form';
 import { Control } from '../../src/types';
-import { Controller } from '../../src';
 
-const App = () => {
-  const {
+type FormValues = {
+  nest: {
+    test: {
+      value: string;
+      nestedArray: {
+        value: string;
+      }[];
+    }[];
+  };
+};
+const ChildComponent = ({
+  index,
+  control,
+}: {
+  control: Control<FormValues>;
+  index: number;
+}) => {
+  const { fields } = useFieldArray<FormValues>({
+    name: `nest.test.${index}.nestedArray` as const,
     control,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({
-    defaultValues: {
-      test: [{ firstName: 'test' }],
-    },
   });
-  const { fields, prepend } = useFieldArray({
-    control,
-    name: 'test',
-  });
-  console.log('errors', errors);
 
   return (
-    <form onSubmit={handleSubmit(() => {})}>
-      {fields.map((field, index) => {
-        return (
-          <div key={field.id}>
-            <Controller
-              control={control}
-              render={({ field }) => <input {...field} />}
-              name={`test.${index}.firstName`}
-              rules={{ required: true }}
-            />
-          </div>
-        );
-      })}
-      <button
-        type="button"
-        onClick={() =>
-          prepend({
-            firstName: '',
-          })
-        }
-      >
-        prepend
-      </button>
-      <button>submit</button>
-    </form>
+    <div>
+      {fields.map((item, i) => (
+        <input
+          key={item.id}
+          {...control.register(
+            `nest.test.${index}.nestedArray.${i}.value` as const,
+          )}
+        />
+      ))}
+    </div>
   );
 };
 
-export default App;
+const Component = () => {
+  const { register, control } = useForm({
+    defaultValues: {
+      nest: {
+        test: [
+          { value: '1', nestedArray: [{ value: '2' }] },
+          { value: '3', nestedArray: [{ value: '4' }] },
+        ],
+      },
+    },
+  });
+  const { fields, remove, append } = useFieldArray({
+    name: 'nest.test',
+    control,
+  });
+
+  return (
+    <div>
+      {fields.map((item, i) => (
+        <div key={item.id}>
+          <input {...register(`nest.test.${i}.value` as const)} />
+
+          <ChildComponent control={control} index={i} />
+
+          <button
+            type={'button'}
+            onClick={() => remove(i)}
+            data-testid={item.value}
+          >
+            remove
+          </button>
+        </div>
+      ))}
+
+      <button type={'button'} onClick={() => append({ value: 'test' })}>
+        append
+      </button>
+    </div>
+  );
+};
+
+export default Component;