--- a/cypress/integration/useFieldArrayNested.ts
+++ b/cypress/integration/useFieldArrayNested.ts
@@ -155,7 +155,7 @@ describe('useFieldArrayNested', () => {
     cy.get('#submit').click();
 
     cy.get('#result').contains(
-      '{"test":[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]}',
+      '{"test":[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]}',
     );
 
     cy.get(`#nest-append-0`).click();
@@ -229,7 +229,7 @@ describe('useFieldArrayNested', () => {
 
     cy.get('#submit').click();
     cy.get('#result').contains(
-      '{"test":[{"firstName":"prepend","keyValue":[{"name":"update"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"append"}],"lastName":"Luo"}]}',
+      '{"test":[{"firstName":"prepend","keyValue":[{"name":"update"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"append"}]}]}',
     );
 
     cy.get('#nest-remove-all-3').click();
@@ -247,7 +247,7 @@ describe('useFieldArrayNested', () => {
 
     cy.get('#submit').click();
     cy.get('#result').contains(
-      '{"test":[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[],"lastName":"Luo"}]}',
+      '{"test":[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[]}]}',
     );
 
     cy.get('#remove').click();