--- a/cypress/integration/useFieldArrayUnregister.ts
+++ b/cypress/integration/useFieldArrayUnregister.ts
@@ -102,7 +102,7 @@ describe('useFieldArrayUnregister', () => {
 
     cy.get('input[name="data.2.name"]').clear().type('bill');
 
-    cy.get('input[name="data.2.conditional"]').should('has.value', 'test');
+    cy.get('input[name="data.2.conditional"]').should('has.value', '');
 
     cy.get('#dirtyFields').should(($state) =>
       expect(JSON.parse($state.text())).to.be.deep.equal({
@@ -129,7 +129,7 @@ describe('useFieldArrayUnregister', () => {
 
     cy.get('#delete1').click();
 
-    cy.get('input[name="data.1.conditional"]').should('has.value', 'test');
+    cy.get('input[name="data.1.conditional"]').should('has.value', '');
 
     cy.get('#submit').click();
 
@@ -137,7 +137,7 @@ describe('useFieldArrayUnregister', () => {
       expect(JSON.parse($state.text())).to.be.deep.equal({
         data: [
           { name: '5' },
-          { name: 'bill', conditional: 'test' },
+          { name: 'bill', conditional: '' },
           { name: '10' },
           { name: 'test1' },
           { name: 'test2' },
@@ -153,7 +153,7 @@ describe('useFieldArrayUnregister', () => {
       expect(JSON.parse($state.text())).to.be.deep.equal({
         data: [
           { name: '5' },
-          { name: 'bill', conditional: 'test' },
+          { name: 'bill', conditional: '' },
           { name: '10' },
           { name: 'test1test' },
           { name: 'test2' },
@@ -169,7 +169,7 @@ describe('useFieldArrayUnregister', () => {
       expect(JSON.parse($state.text())).to.be.deep.equal({
         data: [
           { name: '5' },
-          { name: 'bill', conditional: 'test' },
+          { name: 'bill', conditional: '' },
           { name: '10' },
           { name: 'test2' },
         ],