--- a/cypress/integration/useWatchUseFieldArrayNested.ts
+++ b/cypress/integration/useWatchUseFieldArrayNested.ts
@@ -30,7 +30,7 @@ describe('useWatchUseFieldArrayNested', () => {
     cy.get('#insert').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]',
+      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]',
     );
 
     cy.get(`#nest-append-0`).click();
@@ -40,7 +40,7 @@ describe('useWatchUseFieldArrayNested', () => {
     cy.get(`#nest-move-0`).click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}],"lastName":"Luo"}]',
+      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"1c"},{"name":"append"}]}]',
     );
 
     cy.get('#nest-update-3').click();
@@ -51,20 +51,20 @@ describe('useWatchUseFieldArrayNested', () => {
     );
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"update"},{"name":"append"}],"lastName":"Luo"}]',
+      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"append"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"update"},{"name":"append"}]}]',
     );
 
     cy.get('#nest-update-0').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"update"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"update"},{"name":"append"}],"lastName":"Luo"}]',
+      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"update"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"1a"},{"name":"update"},{"name":"append"}]}]',
     );
 
     cy.get('#nest-remove-3').click();
     cy.get('#nest-remove-3').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"update"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[{"name":"insert"},{"name":"append"}],"lastName":"Luo"}]',
+      '[{"firstName":"prepend","keyValue":[{"name":"insert"},{"name":"prepend"},{"name":"update"}]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[{"name":"insert"},{"name":"append"}]}]',
     );
 
     cy.get('#nest-remove-all-3').click();
@@ -73,7 +73,7 @@ describe('useWatchUseFieldArrayNested', () => {
     cy.get('#nest-remove-all-0').click();
 
     cy.get('#result').contains(
-      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","keyValue":[],"lastName":"Luo"}]',
+      '[{"firstName":"prepend","keyValue":[]},{"firstName":"insert","keyValue":[]},{"firstName":"append","keyValue":[]},{"firstName":"Bill","lastName":"Luo","keyValue":[]}]',
     );
 
     cy.get('#remove').click();