--- a/src/__tests__/logic/unsetEmptyArray.ts
+++ b/src/__tests__/logic/unsetEmptyArray.ts
@@ -1,0 +1,6 @@
+import compact from '../../utils/compact';
+import get from '../../utils/get';
+import unset from '../../utils/unset';
+
+export default <T>(ref: T, name: string) =>
+  !compact(get(ref, name, [])).length && unset(ref, name);