# Improvement

We focus a lot on performance! This is one of the reasons hook form is getting popular, and we would like to push performance even further, we reduced re-render and now it's time to improve the internal when users subscribed to inputs updated. The current design tight user inputs reference with value, however, this also has an impact on deeply nested input value lookup, with the new value store. We can skip that nested value lookup and straightway access the value store. 

This PR has the following improvements:

- Dedicated value store
- Improve perf `getValues`, `watch`, `useWatch`
- enabled register ref input lookup https://github.com/react-hook-form/react-hook-form/issues/6098 
- Disabled inputs value will be updated during `register`
- `setValue` getting uplifted with value update instead rely on `register` inputs
- enable `trigger` for field array and object fields. #2381
- package size reduction **5%**! yay!
- `createFormControl` move logic into its own factory function #5659 credit and idea goes to https://github.com/MHA15
  -  PR for the above work: https://github.com/react-hook-form/react-hook-form/pull/5987

# Added

- register('name', { disabled: boolean })

Currently, the `disabled` attribute is working with a ref callback, this is not an ideal solution. The following prop will guarantee and as a prop solution for toggle `disabled` input.

- `disabled` prop to solve toggle disable for the input value 

```tsx
register('name', { disabled: true })
````

- useWatch({ disabled: boolean })

This new config will allow users to pause/unsubscribe the watch, and you can resume the subscription by toggle the `disable`

```tsx
useWatch({ disabled })
```

- useFormState({ disabled: boolean })

This new config will allow users to pause/unsubscribe formState, and you can resume the subscription by toggle the `disable`

```tsx
useFormState({ disabled })
```

# Result (Dev build)

The following test is tested with my personal laptop below and using chrome latest build with **CPU throttle with x 4**

## 👨‍💻  Testing code: 

### Flat Example: 

```tsx
import * as React from 'react';
import { useForm } from 'react-hook-form';

const Component = () => {
  const { register, watch } = useForm();
  console.log(watch());

  return (
    <div>
      {[...Array(1000).keys()].map((item, i) => (
        <div key={i}>
          <input {...register(`nest${i}`)} />
        </div>
      ))}
    </div>
  );
};
```

### Nested Example: 

```tsx
import * as React from 'react';
import { useForm } from 'react-hook-form';

const Component = () => {
  const { register, watch } = useForm();

  console.log(watch());

  return (
    <div>
      {[...Array(1000).keys()].map((item, i) => {
        const nest = i / 100;
        return (
          <div key={i}>
            <input {...register(`nest${'.test'.repeat(nest)}${i}`)} />
          </div>
        );
      })}
    </div>
  );
};
```

### Deeply Nested Example: (1000 * 2 * 1000)

```tsx
import * as React from 'react';
import { useForm, useWatch } from 'react-hook-form';

const Nest = ({ control }) => {
  useWatch({
    control,
  });
  return null;
};

const Component = () => {
  const { register, watch, control } = useForm();

  console.log(watch());

  return (
    <div>
      {[...Array(1000).keys()].map((item, i) => {
        const nest = i / 100;
        return (
          <div key={i}>
            <input {...register(`nest${'.test'.repeat(nest)}${i}`)} />
            <Nest control={control} />
          </div>
        );
      })}
    </div>
  );
};
```

## 🧾  Outcome:

`7.11.2-beta.0`

### Flat

Average: **18ms** each keystroke to watch input and render the output of watch result

<img width="1440" alt="Screen Shot 2021-07-17 at 11 02 38 am" src="https://user-images.githubusercontent.com/10513364/126020800-11f6d983-af6b-4410-83bc-1005b6c464bc.png">

Mouting:  **470ms** 

<img width="1437" alt="Screen Shot 2021-07-19 at 11 21 05 am" src="https://user-images.githubusercontent.com/10513364/126089424-afa23e40-c6c3-4d4e-b60a-e30e5fe3c144.png">

### Nested

Average: **20ms** each keystroke to watch input and render the output of watch result

<img width="1440" alt="Screen Shot 2021-07-17 at 11 05 06 am" src="https://user-images.githubusercontent.com/10513364/126020849-f6c747c7-f8fb-4bd6-a513-a2f84b927fd1.png">

### Deeply Nested (2000 * 2 * 2000)

Average: **550ms** each keystroke to watch input and render the output of watch result
<img width="1440" alt="Screen Shot 2021-07-17 at 11 37 52 am" src="https://user-images.githubusercontent.com/10513364/126021531-03ea4dd3-e58f-422f-8b33-cb92596d2169.png">

with `createFormControl`

Average: **200ms** each keystroke to watch input and render the output of watch result
<img width="1440" alt="Screen Shot 2021-07-18 at 9 34 06 pm" src="https://user-images.githubusercontent.com/10513364/126065493-cfbd8d28-2d83-484c-85e5-e025b5665e3b.png">


`7.11.1`

### Flat

Average: **40ms** each keystroke to watch input and render the output of watch result

<img width="1440" alt="Screen Shot 2021-07-17 at 10 49 29 am" src="https://user-images.githubusercontent.com/10513364/126020441-df69a275-a7dd-400c-a45f-df385b5a289d.png">

Mouting:  **470ms** 

<img width="1393" alt="mout 7 11 1" src="https://user-images.githubusercontent.com/10513364/126089446-e0d1e29d-e6a5-497f-9442-847219dd2893.png">

### Nested

Average: **43ms** each keystroke to watch input and render the output of watch result

<img width="1308" alt="Screen Shot 2021-07-17 at 11 44 54 am" src="https://user-images.githubusercontent.com/10513364/126021669-8d24e742-20c6-4c10-8f16-ea1e40150576.png">

### Deeply Nested (2000 * 2 * 2000)

Average: **2000ms** each keystroke to watch input and render the output of watch result

<img width="1436" alt="Screen Shot 2021-07-17 at 11 30 38 am" src="https://user-images.githubusercontent.com/10513364/126021432-1c18d76d-2cc2-4d2d-8977-8308fe5e552a.png">

# Conclusion 

Dev Build + Throttle 4x CPU + Chrome + MAC M1 

| Version | Flat (1000) | Nested (1000)  | Deeply Nested (1000 * 2 * 2000) |  Flat (1000) Mount |
|:--- |:---:|:---:|:---:|:---:|
| `7.11.2-beta.0` | 18ms | 20ms | 550ms |  470ms|
| `7.11.2-beta.0` + `createFormControl` | || 200ms (huge improvement) |  470ms|
| `7.11.1` | 40ms | 43ms | 2000ms |  470ms|

The outcome is the new build is faster than the current build. Have a dedicated value store avoided the nested lookup value. The next step will release a beta version and let more users test out this build before release the final minor update.

# Todos

- [x] unit tests coverage
- [x] docs update
- [x] functional tests
- [x] improve `useFieldArray` perf