@Aryaman73 2021-08-05
@bluebill1049 thanks a lot for this PR! 🥳 
 
I noticed a possible bug with the change to `trigger()`. It seems that after upgrading to `7.13.0-next.0`, an error shows up when you pass something to trigger that doesn't exist (e.g `trigger("random-field-that-does-not-exist):`)

![image](https://user-images.githubusercontent.com/53343807/128397132-ba3d467a-d4aa-4d6c-9609-38b426e6c004.png)

Looks like there isn't a catch for if there aren't any fields with that name. I could create an issue for this and code sandbox if necessary, but unless I'm mistaken this should be easy to reproduce in any project. We are relying on useFormContext, if that matters.

Thanks again!