@bluebill1049 2021-08-05
It's better to crash your app at dev before it breaks your prod. `trigger` `name` should be part of your form.